feat: simplify avatar and banner image components by removing unnecessary router handling and restructuring dialog usage
