+++ src/app/(protected)/[username]/(profile)/@image/banner/page.tsx
@@ -1,29 +1,20 @@
 "use client";
 
 import Image from "next/image";
-import { useParams, useRouter } from "next/navigation";
+import { useParams } from "next/navigation";
 import { useQueryClient } from "@tanstack/react-query";
 
 import { type ProfilePageUser } from "@/types/user";
-import { ImageDialog, ImageDialogContent } from "@/components/ui/image-dialog";
+import { ImageDialogContent } from "@/components/ui/image-dialog";
 
 export default function AvatarImagePage() {
-	const router = useRouter();
 	const queryClient = useQueryClient();
 	const { username } = useParams<{ username: string }>();
 	const user = queryClient.getQueryData<ProfilePageUser>([`profile`, username]);
 
-	const handleClose = () => {
-		history.length > 1 ? router.back() : router.push(`/${username}`);
-	};
-
-	if (!user || !user.bannerUrl) {
-		return handleClose();
-	}
-
 	return (
-		<ImageDialog open={true} onOpenChange={handleClose}>
-			<ImageDialogContent className="max-w-auto w-full rounded-none">
+		<ImageDialogContent className="max-w-auto w-full rounded-none">
+			{user?.bannerUrl && (
 				<Image
 					src={user.bannerUrl}
 					width={600}
@@ -31,7 +22,7 @@ export default function AvatarImagePage() {
 					alt={`${user.name} banner`}
 					className="aspect-header w-full object-cover object-center"
 				/>
-			</ImageDialogContent>
-		</ImageDialog>
+			)}
+		</ImageDialogContent>
 	);
 }
